Move shared logic into common files
