+++ src/common/crypto.h
@@ -1,0 +1,11 @@
+//
+// Created by max on 10/16/24.
+//
+
+#ifndef CRYPTO_H
+#define CRYPTO_H
+
+int aes_encrypt();
+int aes_decrypt();
+
+#endif //CRYPTO_H
