+++ src/common/prompt.h
@@ -1,0 +1,6 @@
+#ifndef PROMPT_H
+#define PROMPT_H
+
+void prompt_user();
+
+#endif  // PROMPT_H
